In this part, I used auto layout in order to organise the labels and buttons properly so that they were alliigned equally:
There is also some changes that I had committed but did not push to the GitHub repository: They are below
-
-
-
-
-
-
There were a number of changes that took place. First of all, the reference of MMDrawerController in app delegate was removed since it was causing the app to function incorrectly
I tried many ways to figure out how to get the side draw of the 'Welcome' view controller when the user presses the 'begin' button. This was difficult because originally, i programmed the MMDrawerController so that it works on only one view controller. In this example, I had to use the second view controller as the 'centreviewcontroller'.
Thus, i declared MMdrawercontroller in ViewController( the view controller with the 'begin' button). Rather than setting up the side panel when the app launched (this happened when mmdrawercontroller was launched in app delegate), i set up MMdrawercontroller when the button begin was pressed. I did this by importing App Delegate into the ViewController.m file and copy and pasting the code for MMdrawercontroller from AppDelegate to View Controller:
Appdelegate *appdel = [[AppDelegate sharedapplication] delegate] (Using this had problems, so it was replaced with ---- Appdelegate *appdel = (AppDelegate*)[[AppDelegate sharedapplication] delegate]

the root view controller was then assigned by using :

appdel.window.rootviewcontroller rather than _window.rootviewcontroller