In this section, I have successfuly managed to program the first questions and display the notifications in the form of a UILabel that was added to the thirdviewcontroller scene.  The question answer bank has also received new functions with more arguments. The void functions were removed as I felt the functions here would be useful. These function descriptions are here
-(NSString*) printquestion: (NSString*) testtypefunc :(int) index;

           This function is responsible for printing the initial question. It is called printquestion and it has 2 arguments; testtypefunc (type of test) and index (index of array).  In this example, if the type of test(testtypefunc) is Numerical Reasoning, the resulting question (finalquestion) is accessed from the questionfornumerical at its array index (index). As its the first question, the index is 0. It returns an NSString so the NSString variable being returned is finalquestion (declared in h file)

-(NSString*) printanswerchoices: (NSString*) testtypefunc :(int) index;

This is a similar function to printquestion, only instead of printing questions, it prints the choices. It uses strings from the choicesfornumerical array and fetches the corresponding choice according to index (index). The value returned is an NSStrign called choice (declared in the h file)

-(NSString*) checkanswer: (NSString*) testtypefunc :(NSString*) answer :(int) index;

This is a function that is designated to designated to check if the answer is correct. The checkanswer function returns an NSString variable called result (declared in h file). It uses 3 arguments; the type of test (testtypefunc), the choices provided (answer) and the actual correct answer fetched from an array of specific index (index). First, we need to determine which type of test it is so we know which array to access. Next, we need to determine if the selected answer (answer) is the same as the actual answer fetched from the answersfornumerical array at a specified index. if they are the same, the result(result) will be a string called Correct. Otherwise it will be a string called incorrect.

-----------------------------------------------------------------------

In the thirdviewcontroller file, those functions in the questionanswerBank are called appropriately. For example:

For setting the questions, the questionanswerbank is allocated memory. The printquestion function is used in viewDidLoad and uses @"Numerical Reasoning" and index = 0. When the user clicks on the Numerical Reasoning tab in the tableview controller, the type of test is passed on to the BeginViewController. This type of test is then passes onto the thirdviewcontroller and it is declared as testtype. This testtype variable is the assigned to another variable finaltest.  First we need to determine whether the type of test is Numerical Reasoning. Then the printquestion is called with finaltest set as (will be set as: in this case its set as Numerical Reasoning ) testtypefunc of the function and the index is set to 0 as its the first question we need to access. In this case the testtypefunc is the same as Numerical Reasoning thus the finalquestion is accessed.

Similar steps are carried out with the other functions. The function methods are defined in the m file of QuestionAnswerBank. To understand the function, you simply need to substitute in the arguments assigned in the thirdviewcontroller and consolidate the m file of QuestionAnswerBank.

In the next commit, i will remove some additional bits such as unwanted functions and preform minor adjustments to code.

-(NSString*) nextquestionfunc;  (I am not sure if I am going to do this yet as I am still unsure as to whether we need to display all questions in one view controller or have separate view controller. Once i feel confident with the questionbank function I will make the appropriate decision later on) If i am going to use the same view controller Im going to need some for loops however if I use multiple view controllers, there is more freedom to include a variety of questions such as image questions (image matches etc) so there is a possibility this function might be discarded.