I have been working on trying to solve the problem of the test type and question number not being passed onto the different functions. I have finally figured out the problem. In order to use values outside, we need to synthesise typeoftest and questionnumber and after doing so, the quiz app was working as it should. To think, I spent long hours playing around with functions when I only needed 1 line!
The functions are all over the place  and the code is too long-  I plan to tidy everything up and finish the app very soon. I would like the code to look like the previous commit but with the addition of synthesise and I have determined the code works.

Also i need to create additional arrays for the choices of each questions and I need to finish of the Questions and answers for the General Knowledge and riddles sections