In this part, i have been changing the code so that random question would be generated. I have used an ar4random generator to define the self.questionnumber. There is a problem as some questions might repeat when generating the random number.
In order to get the rand number function working, i had to remove the self.choicenumber as find a pattern between the objectatindex between the question bank and the choices bank. It turns out that if self.questionnumber is 1 (i.e. the second question is called) then the corresponding choices are located at index 4,5,6 and 7. This means that when printing the first choice, we need to print at self.questionnumber * 4. We then increment one each time for the second choice, third choice and fourth choice.

I have also increased the width of the labels, so when we include longer question with bigger answer choices.