In this commit, I have been playing around with the Timerlabel and attempting to pass the timer value onto the resultviewcontroller.  I have managed to do this by referring to the documentation of the MZTimerLabel in GitHub and referring to the demo code that was provided in the zip file. In the previous commit, I have used the UIAlertController in order to give the user the notification on whether to quit the test or not. Since I want the timer value to be used throughout the program, I have declared the MZTimerlabel as *label in the h file and have synthesised it.  Thus in the viewdidLoad, the label is associated with MZTimerlabel, and this is allocated memory and initiated with the timer label. This is a label that has been put on to the storyboard. As MZTimerlabel is a subclass of the UILabel, when the timer begin [label start], it will replace the default self.timerlabel with the clock
When the user finishes the test, I want the user to be notified of the time elapsed. According to the documentation, I used the gettimecounted function from the MZTimerLabel m file using NSTimeInterval. Then the value of the time counted was declared into an NSString with a specific format.  The time elapsed is shown as an alert which is initialised when [alertview show] is carried out. However, this is a deprecated method, thus I have to use the UIAlertcontroller that I used when pressing the quit button.

In addition, there was a problem with choice A where it was not calling the next view controller. This was swiftly fixed by adding in the same code used for the other buttons.