In this commit, I have added some additional features to this app. I animated the label in the welcome page. This was done by downloading a class from Github called RQShinelabel. The original uilabel declarations were removed (*welc and *swipe), and the new labels were given custom classes called RQShineLabel *welcanimate and *swipanimate.  However, I have had to make some changes to the RQShinelabel m file. In the m file, the custom colour was set to white thus this was changed to black.  In the viewDidLoad of the welcome screen, the functions for the animation were called by using 'shine' to animate the label. This was done to both the swipe and welcome label
Additionally, I have been working on the final view controller. There were some problems I initially faced. When I put the button on the UIView of resultsviewcontroller, the button were not being 'selected' or 'pressed' ; they behaved like UILabels. I found that this happened because the confetti function was added as a subview onto the current UIView. Thus I had to add a subview for each of the buttons. This was done by assigning the buttons as a @property type and calling [sefl.view addsubview: (button)]. This was a working part

I have also added IBAction for the button. There is a button for category pressed and try again. When the user presses try again, the user will be able to attempt the test with different questions again. When the user changes category, the welcome view controller (firstViewController) will be called with the side panel activated. The same code used to set up the side panel is used in the category pressed button action (note: this does not require any segues).

I have also added labels in the ResultViewController which will display the time it took to complete the quiz, the score and some improvements depending on the score.