In this commit, I have been implementing the score label and score into the applications. There have been some problems that are related to this variable. When I used an NSLog to print out the score label, I fount out that there was a problem with the score being printed. The answer was being multiplied by 2 when adding 1. This was a very unusual problem, however I have found a fix for this problem. The score variable was declared as a pointer rather than an int type. I have removed the * mark and synthesised score and I have got a working code. I have also added one other label that will display the score answer as I don't want the whole score label to animate; I only want one number to animate every time the user presses the button choice.
I have also added a dedicated view for the result finish text in the ResultviewController. This is because in order for the animation to work, we need to add a subview as the confetti as also been added as the subview.

I have also declared the result finish and improvement labels in the ResultViewController.h file. Each label has been given its own class. The improvement, test type , score and time are all displayed as animations