In this commit, I have successfully implemented the score into the quiz and have managed to pass the value of the time elapsed and the score onto the next view controller (ResultViewController). It took sometime to figure out the code to implement the score. Essentially, every time the user answers a question correctly, the score will increase by 1 and if the user gets an answer wrong the score does not change. However, this is very basic at the moment, and i plan to use and arc 4 random function in order to add the score and deduct the score. In the thirdviewcontroller, if the test type is the same as the test being carried out, the score increases by 1 otherwise the score does not change (shown between line 128 and 142 in thirdviewcontroller.m).
In the segue function, the time passed on is declared as an NSTimeInterval and this has been replaced in the else statements in line 162 of thirdviewcontroller.m.
The time elapsed is passed on by using [label getimecounted] in the segue function and this is equal to passedontime which is declared as NSTimeinterval in ResultViewController . The score is also passed on.

In the Results view Controller, I have added an additional button which will calculate the Final Score. I have imported a ticker score file from Github that will carry out an animation to display the score. I plan to calculate the final score deepening on how long the user took to finish the test. Since there is a confetti present, in order for the button to work, we need to add a subview for the info button so that the button can be pressed. When the button is pressed, the final score is calculated and the improvements will be displayed.