I have been making a number of changes in this commit. First of all I added a separate screen for when the score reaches negative. When this happens the user goes to the failviewcontroller where the user is offered the chance to try again or choose a different category.
Secondly, I have made changes to the types of categories that I am going to include. Instead of including riddles (which i declared on me documentation) I will use true/false questions. This is because there weren't a variety of multiple choice questions available and secondly a riddles category would not be suitable for the theme of the application. This means that in the data model, (implementtypeanddes) I had to change the riddles into True/False Trivia. I had to make this change in the tableviewcontroller file as well.

I main some adjustments in the storyboard in terms of colour assignment and changing the position of different variables such as UIabels.

The Question Bank has all the questions, answers and choices for the True False category.

In the Result View Controller, I have added a ticker label which calculates the final score. However i have used [self performselector afterdelay] functions in order to delay the score calculate functions. Depending on the score, obtained the final score is calculated depending on how high the score in the previous round was.

The score was finally implemented in thridviewcontroller. In the quiz, the user will start of with a score of 10. For every correct answer, the score increases by 10. For every wrong answer, the score decreases by a random number between 0 and 15.

There are different multipliers so the higher the score, the greater the multiplier. When the score is calculated in ResultViewController, the progress hud is displayed in addition.

In each view controller, I have also added code that sets the border for some buttons and labels.